app/requirements.txt
strands-agents>=1.0.0
bedrock-agentcore>=1.0.0
aws-opentelemetry-distro
datasets>=4.5.0
